refactor(mobile): split android build metadata
